INVGEN-42925: support objects paging

+++ WebApplication/Controllers/ProjectsController.cs
@@ -3,7 +3,6 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Threading.Tasks;
-using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
 using WebApplication.Definitions;
@@ -66,6 +65,12 @@ namespace WebApplication.Controllers
         [HttpPost]
         public async Task<ActionResult<ProjectDTO>> CreateProject([FromForm]NewProjectModel projectModel)
         {
+            if (!_userResolver.IsAuthenticated)
+            {
+                _logger.LogError("Attempt to create project for anonymous user");
+                return BadRequest();
+            }
+
             var projectName = Path.GetFileNameWithoutExtension(projectModel.package.FileName);
             var bucket = await _userResolver.GetBucketAsync(true);
 
@@ -92,44 +97,10 @@ namespace WebApplication.Controllers
             ProjectStorage projectStorage = await _userResolver.GetProjectStorageAsync(projectName);
 
             string ossSourceModel = projectStorage.Project.OSSSourceModel;
-            await using (var fileReadStream = System.IO.File.OpenRead(file.Name))
-            {
-                // determine if we need to upload in chunks or in one piece
-                long sizeToUpload = fileReadStream.Length;
-                long chunkMBSize = 5;
-                long chunkSize = chunkMBSize * 1024 * 1024; // 2MB is minimal
-
-                // use chunks for all files greater than chunk size
-                if (sizeToUpload > chunkSize)
-                {
-                    _logger.LogInformation($"Uploading in {chunkMBSize}MB chunks");
-
-                    string sessionId = Guid.NewGuid().ToString();
-                    long begin = 0;
-                    byte[] buffer = new byte[chunkSize];
-                    int bytesRead = 0;
-
-                    while (begin < sizeToUpload-1)
-                    {
-                        bytesRead = await fileReadStream.ReadAsync(buffer, 0, (int)chunkSize);
-                        int memoryStreamSize = sizeToUpload - begin < chunkSize ? (int)(sizeToUpload - begin) : (int)(chunkSize);
-                        using (MemoryStream chunkStream = new MemoryStream(buffer, 0, memoryStreamSize))
-                        {
-                            string contentRange = string.Format($"bytes {begin}-{begin + bytesRead -1}/{sizeToUpload}");
-                            await bucket.UploadChunkAsync(ossSourceModel, contentRange, sessionId, chunkStream);
-                        }
-                        begin += bytesRead;
-                    }
-                }
-                else
-                {
-                    await bucket.UploadObjectAsync(ossSourceModel, fileReadStream);
-                }
-            }
-
-            bool adopted = false;
+            await bucket.SmartUploadAsync(file.Name, ossSourceModel);
 
             // adopt the project
+            bool adopted = false;
             try
             {
                 string signedUrl = await bucket.CreateSignedUrlAsync(ossSourceModel);
@@ -163,45 +134,39 @@ namespace WebApplication.Controllers
         [HttpDelete]
         public async Task<StatusCodeResult> DeleteProjects([FromBody] List<string> projectNameList)
         {
+            if (!_userResolver.IsAuthenticated)
+            {
+                _logger.LogError("Attempt to delete projects for anonymous user");
+                return BadRequest();
+            }
+
             var bucket = await _userResolver.GetBucketAsync(true);
 
-            // delete all oss objects for all provided projects
+            // collect all oss objects for all provided projects
             var tasks = new List<Task>();
 
             foreach (var projectName in projectNameList)
             {
-                var objects = await bucket.GetObjectsAsync($"{ONC.AttributesFolder}-{projectName}");
-                foreach (var objectDetail in objects) {
-                    tasks.Add(bucket.DeleteObjectAsync(objectDetail.ObjectKey));
-                }
+                tasks.Add(bucket.DeleteObjectAsync(Project.ExactOssName(projectName)));
 
-                objects = await bucket.GetObjectsAsync($"{ONC.CacheFolder}-{projectName}");
-                foreach (var objectDetail in objects)
+                foreach (var searchMask in ONC.ProjectMasks(projectName))
                 {
-                    tasks.Add(bucket.DeleteObjectAsync(objectDetail.ObjectKey));
-                }
-
-                objects = await bucket.GetObjectsAsync($"{ONC.DownloadsFolder}-{projectName}");
-                foreach (var objectDetail in objects)
-                {
-                    tasks.Add(bucket.DeleteObjectAsync(objectDetail.ObjectKey));
-                }
-
-                objects = await bucket.GetObjectsAsync($"{ONC.ProjectsFolder}-{projectName}");
-                foreach (var objectDetail in objects)
-                {
-                    tasks.Add(bucket.DeleteObjectAsync(objectDetail.ObjectKey));
+                    var objects = await bucket.GetObjectsAsync(searchMask);
+                    foreach (var objectDetail in objects)
+                    {
+                        tasks.Add(bucket.DeleteObjectAsync(objectDetail.ObjectKey));
+                    }
                 }
             }
 
-            try
+            // delete the OSS objects
+            await Task.WhenAll(tasks);
+            for (var i = 0; i < tasks.Count; i++)
             {
-                Task.WaitAll(tasks.ToArray());
-            }
-            catch (AggregateException e)
-            {
-                _logger.LogError(e, "Failed to delete project(s)");
-                return StatusCode(StatusCodes.Status500InternalServerError);
+                if (tasks[i].IsFaulted)
+                {
+                    _logger.LogError($"Failed to delete file #{i}");
+                }
             }
 
             // delete local cache for all provided projects
